test(cron): move running-state row test beside the capped view suite
